+++ src/core/server/http/router/router.ts
@@ -301,20 +301,10 @@ export class Router implements IRouter {
         return e;
       }
 
-      if (isDataSourceConfigError(e)) {
-        return hapiResponseAdapter.handle(
-          opensearchDashboardsResponseFactory.badRequest({ body: e.message })
-        );
-      }
-
       return hapiResponseAdapter.toInternalError();
     }
   }
 }
-
-const isDataSourceConfigError = (error: any) => {
-  return error.constructor.name === 'DataSourceConfigError' && error.statusCode === 400;
-};
 
 const convertOpenSearchUnauthorized = (
   e: OpenSearchNotAuthorizedError